refactor: moved logging functions into own file
